Move abstract-value-type into own file
