+++ libs/language-server/src/lib/ast/wrappers/value-type/primitive/primitive-value-type.ts
@@ -3,7 +3,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 import { type InternalValueRepresentation } from '../../../expressions/internal-value-representation';
-import { AbstractValueType, type ValueType } from '../value-type';
+import { AbstractValueType } from '../abstract-value-type';
+import { type ValueType } from '../value-type';
 
 export abstract class PrimitiveValueType<
   I extends InternalValueRepresentation = InternalValueRepresentation,
