+++ libs/language-server/src/lib/ast/wrappers/value-type/value-type.ts
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 import { type InternalValueRepresentation } from '../../expressions/internal-value-representation';
-import { type ValuetypeDefinition } from '../../generated/ast';
 
 import { type AtomicValueType } from './atomic-value-type';
 import {
@@ -19,8 +18,6 @@ import {
   type TransformValuetype,
   type ValuetypeAssignmentValuetype,
 } from './primitive';
-
-export type ValueTypeAstNode = ValuetypeDefinition;
 
 export interface VisitableValueType {
   acceptVisitor(visitor: ValueTypeVisitor): void;
@@ -77,63 +74,6 @@ export interface ValueType<
   getName(): string;
 
   equals(target: ValueType): boolean;
-}
-
-export abstract class AbstractValueType<I extends InternalValueRepresentation>
-  implements ValueType<I>
-{
-  abstract acceptVisitor<R>(visitor: ValueTypeVisitor<R>): R;
-
-  isSubtypeOf(other: ValueType): boolean {
-    let othersSupertype = other.getSupertype();
-    while (othersSupertype !== undefined) {
-      if (othersSupertype === this) {
-        return true;
-      }
-      othersSupertype = othersSupertype.getSupertype();
-    }
-    return false;
-  }
-
-  getSupertype(): ValueType | undefined {
-    if (this.hasSupertypeCycle()) {
-      return undefined;
-    }
-    return this.doGetSupertype();
-  }
-
-  protected abstract doGetSupertype(): ValueType | undefined;
-
-  abstract equals(target: ValueType): boolean;
-
-  abstract isAllowedAsRuntimeParameter(): boolean;
-
-  abstract isConvertibleTo(target: ValueType): boolean;
-
-  isReferenceableByUser(): boolean {
-    return false;
-  }
-
-  abstract isInternalValueRepresentation(
-    operandValue: InternalValueRepresentation | undefined,
-  ): operandValue is I;
-
-  abstract getName(): string;
-
-  hasSupertypeCycle(visited: ValueType[] = []): boolean {
-    const cycleDetected = visited.some((v) => v.equals(this));
-    if (cycleDetected) {
-      return true;
-    }
-    visited.push(this);
-
-    const supertype = this.doGetSupertype();
-    if (supertype === undefined) {
-      return false;
-    }
-
-    return supertype.hasSupertypeCycle(visited);
-  }
 }
 
 export abstract class ValueTypeVisitor<R = unknown> {
